Crate secondary_massey

Computes massey products in $\Mod_{C\tau^2}$.

Usage

This computes all Massey products of the form $\langle -, b, a\rangle$, where $a \in \Ext^{*, *}(M, k)$ and $b, (-) \in \Ext^{*, *}(k, k)$. It does not verify that the Massey product is valid, i.e. $a$ and $b$ both lift to $\Mod_{C\tau^2}$ and have trivial product.

Since we must choose $a$ and $b$ to have trivial product, it is necessary to be able to specify the $\tau$ part of them, and not insist that they are standard lifts of the $\Ext$ classes. Thus, the user is first prompted for the $\Ext$ part, then the $\tau$ part of each class. To set a part to zero, supply an empty name. Note that if the bidegree right above the class is empty, the user is not prompted for the $\tau$ part.

Output

This computes the Massey products up to a sign. We write our output in the category $\Mod_{C\tau^2}$, so the format is $\langle a, b, -\rangle$ instead of $\langle -, b, a\rangle$. Brave souls are encouraged to figure out the correct sign for the products.